chore: rename directory dto -> dtos
